requirements.txt
Django==1.8.4
oauthlib==1.0.3
Pillow==2.9.0
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.4.0
wheel==0.24.0
